add create loan and create saving fix some typo mistake

+++ src/pages/Functions/Transfer/Transfer.jsx
@@ -1,13 +1,13 @@
 import {Button, Form, Input, InputNumber, Modal, Radio} from "antd";
 import React, {useState} from "react";
 import {USER_INFO} from "../../../const/key_storage";
-import {getUserInfo, transferMoney} from "../../../api/api_config";
+import {userGetInfo, transactionTransfer} from "../../../api/api_config";
 import "../form.css"
 import AnimatedPage from "../../../utils/AnimatedPage";
 
 const Transfer = (object) => {
     const userInfo = object.object
-    const [receivedPhone, setReceivedPhone] = useState("")
+    const [receiverPhoneNumber, setReceiverPhoneNumber] = useState("")
     const [receiverEmail, setReceiverEmail] = useState("")
     const [money, setMoney] = useState(0.0)
     const [type, setType] = useState("")
@@ -18,19 +18,38 @@ const Transfer = (object) => {
                 <h2 style={{textAlign: "center"}}>Transfer</h2>
                 <Form
                     onFinish={() => {
+                        if (!type) {
+                            Modal.error({title: "Please select send method",})
+                            return
+                        }
+                        if (type === 'phoneNumber' && (!receiverPhoneNumber || !(new RegExp("^(\\+\\d)?\\d{10}$").test(receiverPhoneNumber)))) {
+                            Modal.error({title: "Invalid phone number",})
+                            return
+                        }
+                        if (type === 'email' && (!receiverEmail || !(new RegExp("/^[a-zA-Z0-9]+@(?:[a-zA-Z0-9]+\\.)+[A-Za-z]+$/\n").test(receiverEmail)))) {
+                            Modal.error({title: "Invalid email",})
+                            return
+                        }
+                        if (money <= 0) {
+                            Modal.error({title: "Invalid money to send",})
+                            return
+                        }
                         Modal.confirm(
                             {
-                                title: `Are you sure want to transfer $${money} to ${receivedPhone}`,
+                                title: `Are you sure want to transfer $${money} to ${receiverPhoneNumber}`,
                                 onOk: async () => {
                                     let formData = new FormData();
                                     formData.append("amount", money + "");
-                                    const res = await transferMoney({amount: money, toAccount: receivedPhone})
+                                    const res = await transactionTransfer({
+                                        amount: money,
+                                        toAccount: receiverPhoneNumber
+                                    })
                                     if (!res.ok) Modal.error({
                                         title: "Can not transfer money!",
                                         onOk: () => Modal.destroyAll()
                                     })
                                     else {
-                                        const newUserInfo = await (await getUserInfo()).json()
+                                        const newUserInfo = await (await userGetInfo()).json()
                                         window.localStorage.setItem(USER_INFO, JSON.stringify(newUserInfo.data))
                                         userInfo.set(JSON.stringify(newUserInfo.data))
                                         Modal.success({
@@ -45,29 +64,28 @@ const Transfer = (object) => {
                         )
                     }}>
                     <Form.Item
-                        name="type"
-                        initialValue={type}
-                        rules={[{required: true, message: 'Please pick an item!'}]}>
+                        initialValue={type}>
                         <h3 style={{textAlign: "center"}}>Type of receiver</h3>
-                        <Radio.Group style={{display: "flex", justifyContent: "center", alignItems: "center"}} onChange={(e) => {setType(e.target.value)}}>
+                        <Radio.Group style={{display: "flex", justifyContent: "center", alignItems: "center"}}
+                                     onChange={(e) => {
+                                         setType(e.target.value)
+                                     }}>
                             <Radio.Button value="email">Email</Radio.Button>
                             <Radio.Button value="phoneNumber">Phone number</Radio.Button>
                         </Radio.Group>
                     </Form.Item>
                     {type === 'phoneNumber' &&
-                    <Form.Item
-                        rules={[{required: true, message: 'Please enter receiver account!'}]}>
-                        <h3 style={{textAlign: "center"}}>Receiver's phone</h3>
-                        <Input style={{width: "100%"}} value={receivedPhone} autoComplete={"phone"}
-                               onChange={(e) => setReceivedPhone(e.currentTarget.value)}/>
-                    </Form.Item>}
+                        <Form.Item>
+                            <h3 style={{textAlign: "center"}}>Receiver's phone</h3>
+                            <Input style={{width: "100%"}} value={receiverPhoneNumber} autoComplete={"phone"}
+                                   onChange={(e) => setReceiverPhoneNumber(e.currentTarget.value)}/>
+                        </Form.Item>}
                     {type === 'email' &&
-                    <Form.Item
-                        rules={[{required: true, message: 'Please enter receiver account!'}]}>
-                        <h3 style={{textAlign: "center"}}>Receiver's email</h3>
-                        <Input style={{width: "100%"}} value={receiverEmail} autoComplete={"email"}
-                               onChange={(e) => setReceiverEmail(e.currentTarget.value)}/>
-                    </Form.Item>}
+                        <Form.Item>
+                            <h3 style={{textAlign: "center"}}>Receiver's email</h3>
+                            <Input style={{width: "100%"}} value={receiverEmail} autoComplete={"email"}
+                                   onChange={(e) => setReceiverEmail(e.currentTarget.value)}/>
+                        </Form.Item>}
                     <Form.Item
                         rules={[{required: true, message: 'Please enter your money!'}]}>
                         <h3 style={{textAlign: "center"}}>Money to transfer</h3>
